+++ packages/aws-client/src/types.ts
@@ -1,6 +1,7 @@
 import * as AWS from 'aws-sdk';
 
 import { APIGatewayClient } from '@aws-sdk/client-api-gateway';
+import { CloudWatchEvents } from '@aws-sdk/client-cloudwatch-events';
 import { CloudFormation } from '@aws-sdk/client-cloudformation';
 import { DynamoDBStreamsClient } from '@aws-sdk/client-dynamodb-streams';
 import { DynamoDB, DynamoDBClient } from '@aws-sdk/client-dynamodb';
@@ -10,6 +11,6 @@ import { S3 } from '@aws-sdk/client-s3';
 import { SNS } from '@aws-sdk/client-sns';
 import { SQS } from '@aws-sdk/client-sqs';
 
-export type AWSClientTypes = APIGatewayClient | CloudFormation | DynamoDB
+export type AWSClientTypes = APIGatewayClient | CloudFormation | CloudWatchEvents | DynamoDB
 | DynamoDBClient | DynamoDBStreamsClient | Kinesis | Lambda | S3 | SNS | SQS
 | AWS.Service | AWS.DynamoDB.DocumentClient;