+++ CHANGELOG.md
@@ -41,6 +41,8 @@ terraform values and it will also update the database cluster to use the new con
 
 ### Changed
 
+- **CUMULUS-2888**
+  - Update CloudWatch Events code to AWS SDK v3
 - **CUMULUS-2893**
   - Updated Kinesis code to AWS SDK v3